Implement the `Reader` trait for the `PeekableReader`
The `PeekableReader` can now be used as any other `Reader`.

The implementation makes sure that if the client previously used
the `peek` method, the byte obtained from the underlying reader is
correctly included in the bytes read from the `PeekableReader`.

On the other hand, when the result of the `peek` operation was an
IoError, it will make the result of any read-type operation on
the `PeekableReader` return that same IoError, maintaining the invariant
that the result of the `peek` operation is always the result of the next
`get_byte` `Reader` call.